Add files via upload
Three methods to check and form proper phone number for here in the United States.  The first method is the public entry point.  The second method is private and checks that there are 10 numbers in the string.  The last method places the area code in parenthesis and then places the for formatted phone number after that.